# loop_implementation/api/openapi.yaml
openapi: 3.1.0
info:
  title: Loop Platform API
  version: 0.1.0
  description: |
    Public REST API for the Loop agent platform.

    All endpoints require a Bearer token scoped to a workspace.
    Streaming endpoints additionally support Server-Sent Events (SSE) and WebSocket.
  contact:
    name: Loop Engineering
    email: eng@loop.example
  license:
    name: Apache-2.0
servers:
  - url: https://api.loop.example/v1
    description: Production (na-east — abstract region; concrete cloud may be AWS/Azure/GCP/Alibaba/self-host)
  - url: https://api-eu.loop.example/v1
    description: Production (eu-west — same)
  - url: http://localhost:8080/v1
    description: Local dev
security:
  - BearerAuth: []
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces the caller is a member of
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
    post:
      tags: [Workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
...
  /eval-suites:
    post:
      tags: [Evals]
      summary: Create an eval suite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                cases:
                  type: array
                  items: { $ref: "#/components/schemas/EvalCase" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvalSuite" }
  /eval-suites/{suite_id}/runs:
    parameters: [{ $ref: "#/components/parameters/SuiteId" }]
    post:
      tags: [Evals]
      summary: Start an eval run against an agent version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_version_id]
              properties:
                agent_version_id: { type: string, format: uuid }
                baseline_version_id: { type: string, format: uuid }
      responses:
        "202":
          description: Run queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvalRun" }
...
  /eval-suites/{suite_id}/cases:
    parameters: [{ $ref: "#/components/parameters/SuiteId" }]
    get:
      tags: [Evals]
      summary: List eval cases in suite
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalCaseList"
    post:
      tags: [Evals]
      summary: Add a test case to suite
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvalCase" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvalCase" }
  /workspaces/{workspace_id}/eval-cases/from-resolution:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    post:
      tags: [Evals]
      summary: Save an operator resolution as an eval case
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolutionEvalCaseRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResolutionEvalCaseResponse" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
...
  parameters:
    WorkspaceId:
      in: path
      name: workspace_id
      required: true
      schema: { type: string, format: uuid }
...
    SuiteId:
      in: path
      name: suite_id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        plan: { type: string, enum: [hobby, pro, team, enterprise] }
        region: { $ref: "#/components/schemas/RegionName" }
        created_at: { type: string, format: date-time }
        role: { type: string, enum: [owner, admin, member, viewer] }
    WorkspaceList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    WorkspaceCreate:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        slug: { type: string, pattern: "^[a-z0-9][a-z0-9-]*$" }
        region:
          allOf: [{ $ref: "#/components/schemas/RegionName" }]
          default: na-east
          description: Immutable workspace region pin selected at creation.
...
    RegionName:
      type: string
      enum: [na-east, eu-west]
      description: Abstract Loop region from infra/terraform/regions.yaml.
...
    EvalCase:
      type: object
      required: [name, input, expected, scorers]
      properties:
        id: { type: string, format: uuid }
        suite_id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        name: { type: string }
        input: { type: object }
        expected: { type: object }
        scorers:
          type: array
          items:
            type: object
            properties:
              kind: { type: string, enum: [llm_judge, embedding_sim, regex_match, json_schema, tool_call_assert, latency_le, cost_le, hallucination, toxicity, pii_leak] }
              config: { type: object }
        source: { type: string }
        source_ref: { type: string }
        attachments:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
    EvalCaseList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/EvalCase" }
    ResolutionEvalCaseRequest:
      type: object
      required: [id, title, expectedOutcome, failureReason, linkedTrace, attachments, source]
      properties:
        id: { type: string }
        title: { type: string }
        expectedOutcome: { type: string }
        failureReason: { type: string }
        linkedTrace: { type: string }
        attachments:
          type: array
          items: { type: string }
        source: { type: string, enum: [operator-resolution] }
    ResolutionEvalCaseResponse:
      type: object
      required: [ok, suite_id, case_id, case]
      properties:
        ok: { type: boolean }
        suite_id: { type: string, format: uuid }
        case_id: { type: string, format: uuid }
        case: { $ref: "#/components/schemas/EvalCase" }
    EvalSuite:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        case_count: { type: integer }
    EvalRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        suite_id: { type: string, format: uuid }
        agent_version_id: { type: string, format: uuid }
        status: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
...